docs(capi): add Mandatory Rules section and fix code compliance
Add a "Mandatory Rules" section to CAPI_DESIGN.md consolidating six
rules: type mutability classification, lifecycle functions, thread
safety, naming conventions, complex number FFI convention, and type
definition location.

Fix existing code to comply:
- Move t4a_simplett_f64 struct from simplett.rs to types.rs (Rule 6)
- Add Send + Sync for t4a_simplett_f64 (Rule 3)
- Add is_assigned for t4a_simplett_f64, t4a_simplett_c64 (Rule 2)
- Add is_assigned for t4a_qtci_f64 (Rule 2)
- Add is_assigned for t4a_treetci_f64, t4a_treetci_c64 (Rule 2)
- Document interleaved f64 convention on TreeTciEvalCallbackC64 (Rule 5)
- Update type reference tables in CAPI_DESIGN.md
